requirements.txt
click==7.1.2
numpy==1.19.0
pandas==1.0.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
XlsxWriter==1.2.9
